psypher-events/requirements.txt
flask
flask-sqlalchemy
flask-login
mysql-connector-python
python-dotenv
werkzeug